Close #58: Clean-up registration of job-handlers. This also fixes the broken application start if "io.micronaut:micronaut-management" is added as a dependency.
