+++ micronaut-zeebe-client-feature/src/main/java/info/novatec/micronaut/zeebe/client/feature/ZeebeWorker.java
@@ -16,6 +16,7 @@
 package info.novatec.micronaut.zeebe.client.feature;
 
 import io.micronaut.context.annotation.Executable;
+import io.micronaut.context.annotation.Parallel;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -29,7 +30,8 @@ import java.lang.annotation.Target;
  */
 @Target(value = {ElementType.TYPE, ElementType.METHOD})
 @Documented
-@Executable
+@Executable(processOnStartup = true)
+@Parallel
 public @interface ZeebeWorker {
 
     String type();
